Refactor form preview specific FE functionality to extra module
[MAILPOET-2811]
